requirement.txt
asn1crypto==0.22.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
Flask==0.12.2
future==0.16.0
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
line-bot-sdk==1.3.0
MarkupSafe==1.0
pycparser==2.17
pyOpenSSL==17.0.0
pysolr==3.6.0
requests==2.17.3
six==1.10.0
urllib3==1.21.1
watson-developer-cloud==0.26.0
Werkzeug==0.12.2
